# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.BatchDisableRequest:
    properties:
      post_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - post_ids
    type: object
  handlers.BatchDisableResponse:
    properties:
      disabled:
        type: integer
      failed:
        type: integer
      results:
        items:
          $ref: '#/definitions/handlers.DisableResult'
        type: array
      total:
        type: integer
    type: object
  handlers.BatchToggleRequest:
    properties:
      videos:
        items:
          $ref: '#/definitions/handlers.VideoToggleRequest'
        minItems: 1
        type: array
    required:
    - videos
    type: object
  handlers.BatchToggleResponse:
    properties:
      failed:
        type: integer
      results:
        items:
          $ref: '#/definitions/handlers.ToggleResult'
        type: array
      success:
        type: integer
      total:
        type: integer
    type: object
  handlers.DisableResult:
    properties:
      message:
        type: string
      post_id:
        type: integer
      success:
        type: boolean
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.GetVideoURLRequest:
    properties:
      original_path:
        example: video/2026-05-24/2c61e89ed6_nrSXt1/default_5b41c0/index.m3u8
        type: string
      p720_path:
        example: video/2026-05-24/2c61e89ed6_nrSXt1/720p_5f0e3b/index.m3u8
        type: string
      post_id:
        example: 11434
        type: integer
      quality:
        example: 720p
        type: string
    required:
    - post_id
    type: object
  handlers.GetVideoURLResponse:
    properties:
      duration:
        type: number
      post_id:
        type: integer
      quality:
        type: string
      subtitle_url:
        type: string
      video_url:
        type: string
    type: object
...
  handlers.ToggleResult:
    properties:
      enable:
        type: boolean
      message:
        type: string
      post_id:
        type: integer
      success:
        type: boolean
    type: object
  handlers.VideoListResponse:
    properties:
      all:
        type: boolean
      page:
        type: integer
      posts:
        items: {}
        type: array
      total:
        type: integer
      total_posts:
        type: integer
      total_pages:
        type: integer
      source_total_pages:
        type: integer
    type: object
  handlers.VideoToggleRequest:
    properties:
      enable:
        example: false
        type: boolean
      post_id:
        example: 1
        type: integer
    required:
    - post_id
    type: object
info:
  contact: {}
  description: A Go-Gin based API for video center management. Provides paginated
    or full video lists, batch video toggles, and batch AI/manual review results.
  title: Video Center API
  version: "1.0"
paths:
...
  /proxy/video:
    get:
      description: Proxies video content (m3u8 playlists and ts segments) from upstream
        CDN. Rewrites URLs in m3u8 playlists to route through the proxy.
      parameters:
      - description: Upstream video URL to proxy
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Proxy video content
      tags:
      - videos
...
  /videos:
    get:
      consumes:
      - application/json
      description: Fetch video list from the upstream API with pagination, search,
        and sorting options. Set all=true to aggregate all upstream pages into one
        response.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - default: DESC
        description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - default: false
        description: Load all upstream pages without pagination
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.VideoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get video center list
      tags:
      - videos
  /videos/batch-disable:
    post:
      consumes:
      - application/json
      description: Batch disable multiple videos by calling the upstream video-enable-toggle
        API with enable=false for each post ID
      parameters:
      - description: List of post IDs to disable
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchDisableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BatchDisableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Batch disable videos
      tags:
      - videos
  /videos/batch-toggle:
    post:
      consumes:
      - application/json
      description: Toggle enable/disable status for multiple videos at once via the
        upstream admin API
      parameters:
      - description: List of videos to toggle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BatchToggleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Batch toggle video enable/disable
      tags:
      - videos
  /videos/get-url:
    post:
      consumes:
      - application/json
      description: Fetch the full (non-preview) video URL from the upstream API for
        a given post ID and quality
      parameters:
      - description: Post ID and quality
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GetVideoURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetVideoURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get full video URL
      tags:
      - videos
schemes:
- http
- https
swagger: "2.0"