# handlers/videos.go
package handlers

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"
	url_pkg "net/url"
	"os"
	"strconv"
	"strings"
	"sync"
	"time"

	"github.com/gin-gonic/gin"
)

const (
	defaultBaseURL = "https://v2.yuelk.com"
	defaultAPIKey  = "ef13c2bdf8cd8550ed4c37c323a558c9985d6d928d39a3b53bed864460221d56"
)

func getBaseURL() string {
	if v := os.Getenv("VIDEO_API_BASE_URL"); v != "" {
		return v
	}
	return defaultBaseURL
}

func getAPIKey() string {
	if v := os.Getenv("VIDEO_API_KEY"); v != "" {
		return v
	}
	return defaultAPIKey
}

func getSessionCookie() string {
	return os.Getenv("VIDEO_SESSION_COOKIE")
}

// VideoListResponse represents the response from the video list API
type VideoListResponse struct {
	Posts            []interface{} `json:"posts"`
	Page             int           `json:"page"`
	TotalPages       int           `json:"total_pages"`
	Total            int           `json:"total"`
	TotalPosts       int           `json:"total_posts"`
	All              bool          `json:"all,omitempty"`
	SourceTotalPages int           `json:"source_total_pages,omitempty"`
}

// VideoToggleRequest represents a single video toggle item
type VideoToggleRequest struct {
	PostID int  `json:"post_id" binding:"required" example:"1"`
	Enable bool `json:"enable" example:"false"`
}

// BatchToggleRequest represents the request body for batch toggle
type BatchToggleRequest struct {
	Videos []VideoToggleRequest `json:"videos" binding:"required,min=1"`
}

// ToggleResult represents the result of a single toggle operation
type ToggleResult struct {
	PostID  int    `json:"post_id"`
	Enable  bool   `json:"enable"`
	Success bool   `json:"success"`
	Message string `json:"message"`
}

// BatchToggleResponse represents the response for batch toggle
type BatchToggleResponse struct {
	Success int            `json:"success"`
	Failed  int            `json:"failed"`
	Total   int            `json:"total"`
	Results []ToggleResult `json:"results"`
}

// GetVideoURLRequest represents the request for getting a full video URL
type GetVideoURLRequest struct {
	PostID       int    `json:"post_id" binding:"required" example:"11434"`
	Quality      string `json:"quality" example:"720p"`
	P720Path     string `json:"p720_path" example:"video/2026-05-24/2c61e89ed6_nrSXt1/720p_5f0e3b/index.m3u8"`
	OriginalPath string `json:"original_path" example:"video/2026-05-24/2c61e89ed6_nrSXt1/default_5b41c0/index.m3u8"`
}

// GetVideoURLResponse represents the response from the get-video-url API
type GetVideoURLResponse struct {
	VideoURL    string  `json:"video_url"`
	SubtitleURL string  `json:"subtitle_url"`
	Quality     string  `json:"quality"`
	PostID      int     `json:"post_id"`
	Duration    float64 `json:"duration"`
}

// GetVideoURL godoc
// @Summary Get full video URL
// @Description Fetch the full (non-preview) video URL from the upstream API for a given post ID and quality
// @Tags videos
// @Accept json
// @Produce json
// @Param request body GetVideoURLRequest true "Post ID and quality"
// @Success 200 {object} GetVideoURLResponse
// @Failure 400 {object} ErrorResponse
// @Failure 502 {object} ErrorResponse
// @Router /videos/get-url [post]
func GetVideoURL(c *gin.Context) {
	var req GetVideoURLRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: "invalid request: " + err.Error()})
		return
	}

	if req.Quality == "" {
		req.Quality = "720p"
	}

	pathValues := map[string]string{
		"original_path": req.OriginalPath,
		"p720_path":     req.P720Path,
	}
	for _, key := range videoPathPreference(req.Quality) {
		if videoURL := buildVideoURLFromPath(pathValues[key]); videoURL != "" {
			c.JSON(http.StatusOK, GetVideoURLResponse{
				VideoURL: makeProxyURL(videoURL),
				Quality:  req.Quality,
				PostID:   req.PostID,
			})
			return
		}
	}

	payload, _ := json.Marshal(map[string]interface{}{
		"post_id": req.PostID,
		"quality": req.Quality,
	})

	url := fmt.Sprintf("%s/pyvideo2/get-video-url", getBaseURL())
	client := &http.Client{Timeout: 15 * time.Second}
	httpReq, err := http.NewRequestWithContext(c.Request.Context(), http.MethodPost, url, bytes.NewReader(payload))
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "failed to create request"})
		return
	}
	httpReq.Header.Set("Content-Type", "application/json")
	httpReq.Header.Set("Accept", "application/json")
	httpReq.Header.Set("X-API-KEY", getAPIKey())

	resp, err := client.Do(httpReq)
	if err != nil {
		c.JSON(http.StatusBadGateway, ErrorResponse{Error: "failed to fetch video URL from upstream"})
		return
	}
	defer resp.Body.Close()

	body, err := io.ReadAll(resp.Body)
	if err != nil {
		c.JSON(http.StatusBadGateway, ErrorResponse{Error: "failed to read upstream response"})
		return
	}

	if resp.StatusCode != http.StatusOK {
		c.JSON(resp.StatusCode, ErrorResponse{Error: fmt.Sprintf("upstream returned %d: %s", resp.StatusCode, string(body))})
		return
	}

	// Parse the upstream response and rewrite the video URL
	var upstream map[string]interface{}
	if err := json.Unmarshal(body, &upstream); err != nil {
		c.Data(resp.StatusCode, "application/json", body)
		return
	}

	if videoURL, ok := upstream["video_url"].(string); ok && videoURL != "" {
		upstream["video_url"] = makeProxyURL(replaceVideoHost(videoURL))
	}

	rewritten, err := json.Marshal(upstream)
	if err != nil {
		c.Data(resp.StatusCode, "application/json", body)
		return
	}
	c.Data(http.StatusOK, "application/json", rewritten)
}

// fetchVideoURL fetches the video URL for a given post ID and quality from the
// upstream API. Returns the raw video URL or empty string on failure.
func fetchVideoURL(ctx context.Context, postID int, quality string) string {
	payload, _ := json.Marshal(map[string]interface{}{
		"post_id": postID,
		"quality": quality,
	})

	u := fmt.Sprintf("%s/pyvideo2/get-video-url", getBaseURL())
	client := &http.Client{Timeout: 10 * time.Second}
	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u, bytes.NewReader(payload))
	if err != nil {
		return ""
	}
	req.Header.Set("Content-Type", "application/json")
	req.Header.Set("Accept", "application/json")
	req.Header.Set("X-API-KEY", getAPIKey())

	resp, err := client.Do(req)
	if err != nil {
		return ""
	}
	defer resp.Body.Close()

	if resp.StatusCode != http.StatusOK {
		return ""
	}

	body, err := io.ReadAll(resp.Body)
	if err != nil {
		return ""
	}

	var result map[string]interface{}
	if err := json.Unmarshal(body, &result); err != nil {
		return ""
	}

	if videoURL, ok := result["video_url"].(string); ok {
		return videoURL
	}
	return ""
}

// transformV2Response converts the v2 upstream API response into the flat
// format expected by the frontend and the rest of the pipeline (enrichment,
// URL rewriting). The v2 envelope looks like:
//
//	{ "data": { "videos": [...] }, "count": { "videos": N }, ... }
//
// We transform it into:
//
//	{ "posts": [...], "total_posts": N, "total_pages": P, "page": X }
//
// Each video object is also normalised: post_id → id, first_image_path →
// first_image (full URL), preview_path → preview_video_url (full URL).
func transformV2Response(body []byte, page, perPage int) []byte {
	var envelope map[string]interface{}
	if err := json.Unmarshal(body, &envelope); err != nil {
		return body
	}

	// Detect v2 envelope by the presence of "data.videos".
	dataObj, _ := envelope["data"].(map[string]interface{})
	if dataObj == nil {
		return body // not a v2 response – pass through
	}
	videos, ok := dataObj["videos"].([]interface{})
	if !ok {
		return body
	}

	// Total count lives in count.videos
	totalCount := 0
	if countObj, ok := envelope["count"].(map[string]interface{}); ok {
		if v, ok := countObj["videos"].(float64); ok {
			totalCount = int(v)
		}
	}

	totalPages := 1
	if perPage > 0 && totalCount > 0 {
		totalPages = (totalCount + perPage - 1) / perPage
	}

	playBase := getVideoPlayBaseURL()

	// Normalise each video object
	for _, v := range videos {
		vm, ok := v.(map[string]interface{})
		if !ok {
			continue
		}
		// post_id → id (keep post_id too for backward compat)
		if pid, ok := vm["post_id"].(float64); ok {
			vm["id"] = pid
		}
		// Build full preview_video_url from preview_path
		if pp, ok := vm["preview_path"].(string); ok && pp != "" {
			vm["preview_video_url"] = playBase + "/" + pp
		}
		// Build full first_image URL from first_image_path
		if fip, ok := vm["first_image_path"].(string); ok && fip != "" {
			vm["first_image"] = fmt.Sprintf("%s/pima/wp-content/uploads/%s", getBaseURL(), fip)
		}
	}

	result := map[string]interface{}{
		"posts":       videos,
		"total_posts": totalCount,
		"total_pages": totalPages,
		"page":        page,
	}

	out, err := json.Marshal(result)
	if err != nil {
		return body
	}
	return out
}

// enrichWithVideoURLs fetches 720p video URLs for each post in the upstream
// response and adds them as "video_url" fields. The fetches run in parallel
// with a bounded timeout so the video list response is not overly delayed.
func enrichWithVideoURLs(ctx context.Context, body []byte) []byte {
	var data map[string]interface{}
	if err := json.Unmarshal(body, &data); err != nil {
		return body
	}

	posts, ok := data["posts"].([]interface{})
	if !ok || len(posts) == 0 {
		return body
	}

	// Use a derived context with a timeout so enrichment doesn't block forever
	enrichCtx, cancel := context.WithTimeout(ctx, 15*time.Second)
	defer cancel()

	var wg sync.WaitGroup
	for _, post := range posts {
		postMap, ok := post.(map[string]interface{})
		if !ok {
			continue
		}
		postID, ok := postMap["id"].(float64) // JSON numbers are float64
		if !ok {
			continue
		}
		if buildVideoURLFromPost(postMap) != "" {
			continue
		}

		wg.Add(1)
		// Each goroutine receives its own unique postMap; no shared map writes.
		go func(pm map[string]interface{}, id int) {
			defer wg.Done()
			videoURL := fetchVideoURL(enrichCtx, id, "720p")
			if videoURL != "" {
				pm["video_url"] = videoURL
			}
		}(postMap, int(postID))
	}
	wg.Wait()

	enriched, err := json.Marshal(data)
	if err != nil {
		return body
	}
	return enriched
}

// BatchDisableRequest represents the request body for batch disable
type BatchDisableRequest struct {
	PostIDs []int `json:"post_ids" binding:"required,min=1" example:"1,2,3"`
}

// DisableResult represents the result of a single disable operation
type DisableResult struct {
	PostID  int    `json:"post_id"`
	Success bool   `json:"success"`
	Message string `json:"message"`
}

// BatchDisableResponse represents the response for batch disable
type BatchDisableResponse struct {
	Disabled int             `json:"disabled"`
	Failed   int             `json:"failed"`
	Total    int             `json:"total"`
	Results  []DisableResult `json:"results"`
}

// ErrorResponse represents an error response
type ErrorResponse struct {
	Error string `json:"error"`
}

// GetVideos godoc
// @Summary Get video center list
// @Description Fetch video list from the upstream API with pagination, search, and sorting options. Set all=true to aggregate all upstream pages into one response.
// @Tags videos
// @Accept json
// @Produce json
// @Param page query int false "Page number" default(1) minimum(1)
// @Param per_page query int false "Items per page" default(20) minimum(1) maximum(100)
// @Param search query string false "Search keyword"
// @Param order query string false "Sort order" Enums(ASC, DESC) default(DESC)
// @Param all query bool false "Load all upstream pages without pagination" default(false)
// @Success 200 {object} VideoListResponse
// @Failure 400 {object} ErrorResponse
// @Failure 502 {object} ErrorResponse
// @Router /videos [get]
func GetVideos(c *gin.Context) {
	page, _ := strconv.Atoi(c.DefaultQuery("page", "1"))
	if page < 1 {
		page = 1
	}

	perPage, _ := strconv.Atoi(c.DefaultQuery("per_page", "20"))
	if perPage < 1 {
		perPage = 20
	}
	if perPage > 100 {
		perPage = 100
	}

	search := c.Query("search")
	order := strings.ToUpper(c.DefaultQuery("order", "DESC"))
	if order != "ASC" && order != "DESC" {
		order = "DESC"
	}

	loadAll := strings.EqualFold(c.DefaultQuery("all", "false"), "true") || c.Query("all") == "1"
	if loadAll && c.Query("per_page") == "" {
		perPage = 100
	}
	var transformed []byte
	if loadAll {
		var err error
		transformed, err = fetchAllVideoPages(c.Request.Context(), page, perPage, search, order)
		if err != nil {
			if upstreamErr, ok := err.(*upstreamVideoListError); ok {
				c.Data(upstreamErr.StatusCode, "application/json", upstreamErr.Body)
				return
			}
			c.JSON(http.StatusBadGateway, ErrorResponse{Error: err.Error()})
			return
		}
	} else {
		body, statusCode, err := fetchVideoListPage(c.Request.Context(), page, perPage, search, order)
		if err != nil {
			c.JSON(http.StatusBadGateway, ErrorResponse{Error: err.Error()})
			return
		}
		if statusCode != http.StatusOK {
			c.Data(statusCode, "application/json", body)
			return
		}

		// Transform v2 response envelope into the flat format expected downstream.
		transformed = transformV2Response(body, page, perPage)
	}

	// Enrich posts with 720p video URLs fetched in parallel from upstream
	enriched := enrichWithVideoURLs(c.Request.Context(), transformed)

	// Rewrite video URLs to go through our proxy so browsers can access them
	rewritten := rewriteVideoURLs(enriched)
	c.Data(http.StatusOK, "application/json", rewritten)
}

func disableVideoUpstream(ctx context.Context, postID int) (bool, string) {
	payload, _ := json.Marshal(map[string]interface{}{
		"post_id": postID,
		"enable":  false,
	})

	url := fmt.Sprintf("%s/pyvideo2/api/admin/video-enable-toggle", getBaseURL())
	httpReq, err := http.NewRequestWithContext(ctx, http.MethodPost, url, bytes.NewReader(payload))
	if err != nil {
		return false, "failed to create request"
	}
	httpReq.Header.Set("Content-Type", "application/json")
	httpReq.Header.Set("Accept", "application/json")
	httpReq.Header.Set("X-API-KEY", getAPIKey())

	client := &http.Client{Timeout: 15 * time.Second}
	resp, err := client.Do(httpReq)
	if err != nil {
		return false, "failed to call upstream API"
	}
	defer resp.Body.Close()

	body, _ := io.ReadAll(resp.Body)
	if resp.StatusCode == http.StatusOK {
		return true, "disabled"
	}
	return false, fmt.Sprintf("upstream returned %d: %s", resp.StatusCode, string(body))
}

const maxVideoAggregationPages = 1000

type upstreamVideoListError struct {
	StatusCode int
	Body       []byte
}

func (e *upstreamVideoListError) Error() string {
	return fmt.Sprintf("upstream returned %d", e.StatusCode)
}

func fetchVideoListPage(ctx context.Context, page, perPage int, search, order string) ([]byte, int, error) {
	u := fmt.Sprintf("%s/pyvideo2/api/videos/manage?page=%d&limit=%d&sort_by=created_at&sort_order=%s",
		getBaseURL(), page, perPage, order)
	if search != "" {
		u += "&search=" + url_pkg.QueryEscape(search)
	}

	client := &http.Client{Timeout: 15 * time.Second}
	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u, nil)
	if err != nil {
		return nil, 0, fmt.Errorf("failed to create request")
	}
	req.Header.Set("Accept", "application/json")
	req.Header.Set("X-API-KEY", getAPIKey())

	// The /pyvideo2/api/videos/manage endpoint requires session cookie authentication.
	if sessionCookie := getSessionCookie(); sessionCookie != "" {
		req.AddCookie(&http.Cookie{Name: "pyvideo2_session", Value: sessionCookie})
	}

	resp, err := client.Do(req)
	if err != nil {
		return nil, 0, fmt.Errorf("failed to fetch data from upstream")
	}
	defer resp.Body.Close()

	body, err := io.ReadAll(resp.Body)
	if err != nil {
		return nil, resp.StatusCode, fmt.Errorf("failed to read upstream response")
	}
	return body, resp.StatusCode, nil
}

func parseVideoListPayload(body []byte, page, perPage int) (map[string]interface{}, bool) {
	transformed := transformV2Response(body, page, perPage)
	var payload map[string]interface{}
	if err := json.Unmarshal(transformed, &payload); err != nil {
		return nil, false
	}
	if _, ok := payload["posts"].([]interface{}); !ok {
		return nil, false
	}
	return payload, true
}

func intFromPayload(payload map[string]interface{}, key string, fallback int) int {
	if value, ok := payload[key].(float64); ok {
		return int(value)
	}
	return fallback
}

func videoPayloadKey(video interface{}) string {
	videoMap, ok := video.(map[string]interface{})
	if !ok {
		return ""
	}
	if id, ok := videoMap["id"]; ok {
		return fmt.Sprint(id)
	}
	if id, ok := videoMap["post_id"]; ok {
		return fmt.Sprint(id)
	}
	return ""
}

func fetchAllVideoPages(ctx context.Context, firstPage, perPage int, search, order string) ([]byte, error) {
	// all=true is intentionally independent of the requested page.
	firstPage = 1

	firstBody, statusCode, err := fetchVideoListPage(ctx, firstPage, perPage, search, order)
	if err != nil {
		return nil, err
	}
	if statusCode != http.StatusOK {
		return nil, &upstreamVideoListError{StatusCode: statusCode, Body: firstBody}
	}

	firstPayload, ok := parseVideoListPayload(firstBody, firstPage, perPage)
	if !ok {
		// Preserve compatibility with non-standard upstream responses.
		return transformV2Response(firstBody, firstPage, perPage), nil
	}

	allPosts := make([]interface{}, 0)
	seen := make(map[string]struct{})
	appendPosts := func(payload map[string]interface{}) {
		posts, _ := payload["posts"].([]interface{})
		for _, post := range posts {
			key := videoPayloadKey(post)
			if key != "" {
				if _, exists := seen[key]; exists {
					continue
				}
				seen[key] = struct{}{}
			}
			allPosts = append(allPosts, post)
		}
	}
	appendPosts(firstPayload)

	sourceTotalPages := intFromPayload(firstPayload, "total_pages", firstPage)
	if sourceTotalPages < firstPage {
		sourceTotalPages = firstPage
	}
	if sourceTotalPages-firstPage > maxVideoAggregationPages {
		return nil, fmt.Errorf("too many upstream pages to aggregate")
	}

	for page := firstPage + 1; page <= sourceTotalPages; page++ {
		body, status, fetchErr := fetchVideoListPage(ctx, page, perPage, search, order)
		if fetchErr != nil {
			return nil, fetchErr
		}
		if status != http.StatusOK {
			return nil, &upstreamVideoListError{StatusCode: status, Body: body}
		}
		payload, valid := parseVideoListPayload(body, page, perPage)
		if !valid {
			return nil, fmt.Errorf("invalid video list response from upstream on page %d", page)
		}
		appendPosts(payload)
	}

	totalPosts := intFromPayload(firstPayload, "total_posts", intFromPayload(firstPayload, "total", len(allPosts)))
	if totalPosts < len(allPosts) {
		totalPosts = len(allPosts)
	}
	result := map[string]interface{}{
		"posts":              allPosts,
		"page":               1,
		"total_pages":        1,
		"total_posts":        totalPosts,
		"total":              totalPosts,
		"all":                true,
		"source_total_pages": sourceTotalPages,
	}
	resultBody, err := json.Marshal(result)
	if err != nil {
		return nil, fmt.Errorf("failed to build aggregated video list")
	}
	return resultBody, nil
}

// BatchToggleVideos godoc
// @Summary Batch toggle video enable/disable
// @Description Toggle enable/disable status for multiple videos at once via the upstream admin API
// @Tags videos
// @Accept json
// @Produce json
// @Param request body BatchToggleRequest true "List of videos to toggle"
// @Success 200 {object} BatchToggleResponse
// @Failure 400 {object} ErrorResponse
// @Failure 502 {object} ErrorResponse
// @Router /videos/batch-toggle [post]
func BatchToggleVideos(c *gin.Context) {
	var req BatchToggleRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: "invalid request: " + err.Error()})
		return
	}

	client := &http.Client{Timeout: 15 * time.Second}
	results := make([]ToggleResult, 0, len(req.Videos))
	successCount := 0
	failedCount := 0

	for _, video := range req.Videos {
		result := ToggleResult{
			PostID: video.PostID,
			Enable: video.Enable,
		}

		payload, _ := json.Marshal(map[string]interface{}{
			"post_id": video.PostID,
			"enable":  video.Enable,
		})

		url := fmt.Sprintf("%s/pyvideo2/api/admin/video-enable-toggle", getBaseURL())
		httpReq, err := http.NewRequestWithContext(c.Request.Context(), http.MethodPost, url, bytes.NewReader(payload))
		if err != nil {
			result.Success = false
			result.Message = "failed to create request"
			failedCount++
			results = append(results, result)
			continue
		}
		httpReq.Header.Set("Content-Type", "application/json")
		httpReq.Header.Set("Accept", "application/json")
		httpReq.Header.Set("X-API-KEY", getAPIKey())

		resp, err := client.Do(httpReq)
		if err != nil {
			result.Success = false
			result.Message = "failed to call upstream API"
			failedCount++
			results = append(results, result)
			continue
		}

		body, _ := io.ReadAll(resp.Body)
		resp.Body.Close()

		if resp.StatusCode == http.StatusOK {
			result.Success = true
			result.Message = "ok"
			successCount++
		} else {
			result.Success = false
			result.Message = fmt.Sprintf("upstream returned %d: %s", resp.StatusCode, string(body))
			failedCount++
		}

		results = append(results, result)
	}

	c.JSON(http.StatusOK, BatchToggleResponse{
		Success: successCount,
		Failed:  failedCount,
		Total:   len(req.Videos),
		Results: results,
	})
}

// BatchDisableVideos godoc
// @Summary Batch disable videos
// @Description Batch disable multiple videos by calling the upstream video-enable-toggle API with enable=false for each post ID
// @Tags videos
// @Accept json
// @Produce json
// @Param request body BatchDisableRequest true "List of post IDs to disable"
// @Success 200 {object} BatchDisableResponse
// @Failure 400 {object} ErrorResponse
// @Failure 502 {object} ErrorResponse
// @Router /videos/batch-disable [post]
func BatchDisableVideos(c *gin.Context) {
	var req BatchDisableRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: "invalid request: " + err.Error()})
		return
	}

	results := make([]DisableResult, 0, len(req.PostIDs))
	disabledCount := 0
	failedCount := 0

	for _, postID := range req.PostIDs {
		result := DisableResult{
			PostID: postID,
		}

		result.Success, result.Message = disableVideoUpstream(c.Request.Context(), postID)
		if result.Success {
			disabledCount++
		} else {
			failedCount++
		}

		results = append(results, result)
	}

	c.JSON(http.StatusOK, BatchDisableResponse{
		Disabled: disabledCount,
		Failed:   failedCount,
		Total:    len(req.PostIDs),
		Results:  results,
	})
}